refactor: OpenRouter 클라이언트/응답 공통화 (리뷰 반영)
- common/infra/openrouter 패키지로 OpenRouterChatResponse, OpenRouterWebClientConfig 이동
- OpenRouterChatCaller 도입 — /chat/completions 호출 + retry/isRetryable 스펙 공통화
- NlOpenRouterClient, OpenRouterMaskingClient가 공통 caller/response 사용
- OpenRouterMaskingResponse 제거
- WebClient 빈 조건: nl.provider 또는 masking.provider 중 하나라도 openrouter면 활성화
