+++ src/main/java/com/sports/server/common/infra/openrouter/OpenRouterWebClientConfig.java
@@ -1,14 +1,16 @@
-package com.sports.server.command.nl.infra;
+package com.sports.server.common.infra.openrouter;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Configuration
-@ConditionalOnProperty(name = "nl.provider", havingValue = "openrouter")
+@ConditionalOnExpression(
+        "'${nl.provider:}'.equals('openrouter') or '${masking.provider:}'.equals('openrouter')"
+)
 public class OpenRouterWebClientConfig {
 
     @Value("${openrouter.api.base-url:https://openrouter.ai/api/v1}")
@@ -24,5 +26,10 @@ public class OpenRouterWebClientConfig {
                 .defaultHeader("Authorization", "Bearer " + apiKey)
                 .defaultHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
                 .build();
+    }
+
+    @Bean
+    public OpenRouterChatCaller openRouterChatCaller(WebClient openRouterWebClient) {
+        return new OpenRouterChatCaller(openRouterWebClient);
     }
 }
